requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
babel==2.17.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
Flask==3.0.0
Flask-Babel==2.0.0
Flask-Limiter==3.8.0
flask-talisman==1.1.0
Flask-WTF==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pycparser==2.22
Pygments==2.19.2
python-dotenv==1.0.1
pytz==2025.2
redis==6.4.0
requests==2.32.4
rich==13.9.4
stripe==12.3.0
typing_extensions==4.14.1
urllib3==2.5.0
waitress==3.0.0
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
